Split inlineEditForm.js into 2 modules, and add a "simple" version
The simple version is intended to be used for forms that want the
edit/save/cancel behaviour but want to handle form serialization and
error reporting in a custom way.
