+++ opentreemap/treemap/js/src/inlineEditForm.js
@@ -8,35 +8,11 @@ var $ = require('jquery'),
     _ = require('lodash'),
     moment = require('moment'),
     FH = require('treemap/fieldHelpers'),
-    getDatum = require('treemap/otmTypeahead').getDatum,
     console = require('console-browserify'),
+    editableForm = require('treemap/editableForm'),
 
-    eventsLandingInEditMode = ['edit:start', 'save:start', 'save:error'],
+    eventsLandingInEditMode = [editableForm.editStartAction, 'save:start', 'save:error'],
     eventsLandingInDisplayMode = ['idle', 'save:ok', 'cancel'];
-
-// Placed onto the jquery object
-require('bootstrap-datepicker');
-
-// Boolean fields values are provided as "True" and "False"
-// from the server-side template tags as well as in this module.
-// In order to provide custom values for these fields, this function
-// can be run after writing a value to the boolean field, it will
-// comb through the provided data attributes to see if custom text
-// is provided.
-//
-// To make a field/element function with customizable boolean labels:
-// * specify the data-bool-true-text attribute on the element
-// * specify the data-bool-false-text attribute on the element
-function getBooleanFieldText (boolField, boolText) {
-    var $boolField = $(boolField),
-        attributes = {True: 'data-bool-true-text',
-                      False: 'data-bool-false-text'},
-        attribute = attributes[boolText];
-
-    // .is() is the recommended way of doing 'hasattr'
-    return $boolField.is("[" + attribute + "]") ?
-        $boolField.attr(attribute) : boolText;
-}
 
 exports.init = function(options) {
     var updateUrl = options.updateUrl,
@@ -51,11 +27,13 @@ exports.init = function(options) {
         errorCallback = options.errorCallback || $.noop,
         onSaveBefore = options.onSaveBefore || _.identity,
         onSaveAfter = options.onSaveAfter || _.identity,
-        editStream = $edit.asEventStream('click').map('edit:start'),
+        editStream = $edit.asEventStream('click').map(editableForm.editStartAction),
         saveStream = (options.saveStream || $save.asEventStream('click')).map('save:start'),
         externalCancelStream = BU.triggeredObjectStream('cancel'),
         cancelStream = $cancel.asEventStream('click').map('cancel'),
         actionStream = new Bacon.Bus(),
+
+        editForm = editableForm.init(options),
 
         logError = function(error) {
             console.error("Error uploading to " + updateUrl, error);
@@ -98,101 +76,6 @@ exports.init = function(options) {
             $("table[data-udf-id] .headerrow").css('display', '');
 
             $("table[data-udf-id] .placeholder").css('display', 'none');
-        },
-
-        displayValuesToTypeahead = function() {
-            $('[data-typeahead-restore]').each(function(index, el) {
-                var field = $(el).attr('data-typeahead-restore');
-                if (field) {
-                    $('input[name="' + field + '"]').trigger('restore', $(el).val());
-                }
-            });
-        },
-
-        displayValuesToFormFields = function() {
-            $(displayFields).each(function(index, el) {
-                var $el = $(el),
-                    field = $el.attr('data-field'),
-                    value = $el.attr('data-value'),
-                    $input;
-
-                if (field && $el.is('[data-value]')) {
-                    $input = FH.getSerializableField($(editFields), field);
-                    if ($input.is('[type="checkbox"]')) {
-                        $input.prop('checked', value == "True");
-                    }
-                    else if ($input.is('[data-date-format]')) {
-                        FH.applyDateToDatepicker($input, value);
-                    } else {
-                        $input.val(value);
-                    }
-                }
-            });
-            displayValuesToTypeahead();
-        },
-
-        typeaheadToDisplayValues = function() {
-            $('[data-typeahead-input]').each(function(index, el) {
-                var datum = getDatum($(el)),
-                    field = $(el).attr('data-typeahead-input');
-                if (typeof datum != "undefined") {
-                    $('[data-typeahead-restore="' + field + '"]').each(function(index, el) {
-                        $(el).val(datum[$(el).attr('data-datum')]);
-                    });
-                    $('[data-typeahead="' + field + '"]').each(function(index, el) {
-                        $(el).html(datum[$(el).attr('data-datum')]);
-                    });
-                }
-            });
-        },
-
-        formFieldsToDisplayValues = function() {
-            $(editFields).each(function(index, el){
-                var field = $(el).attr('data-field'),
-                    $input, value, display, digits, units,
-                    displayValue;
-
-                // if the edit field has a data-field property,
-                // look for a corresponding display value and if
-                // found, populate the display value
-                if ($(el).is('[data-field]')) {
-                    display = FH.getField($(displayFields), field);
-
-                    if ($(display).is('[data-value]')) {
-                        $input = FH.getSerializableField($(editFields), field);
-                        if ($input.is('[type="checkbox"]')) {
-                            value = $input.is(':checked') ? "True" : "False";
-                        } else if ($input.is('[data-date-format]')) {
-                            value = FH.getTimestampFromDatepicker($input);
-                        } else {
-                            value = $input.val();
-                        }
-
-                        $(display).attr('data-value', value);
-                        displayValue = value;
-
-                        if ($input.is('select')) {
-                            // Use dropdown text (not value) as display value
-                            displayValue = $input.find('option:selected').text();
-                        } else if ($input.is('[type="checkbox"]')) {
-                            displayValue = getBooleanFieldText(display, value);
-                        } else if (value && $input.is('[data-date-format]')) {
-                            displayValue = $input.val();
-                        } else if (value) {
-                            digits = $(display).data('digits');
-                            if (digits) {
-                                displayValue = parseFloat(value).toFixed(digits);
-                            }
-                            units = $(display).data('units');
-                            if (units) {
-                                displayValue = value + ' ' + units;
-                            }
-                        }
-                        $(display).text(displayValue);
-                    }
-                }
-            });
-            typeaheadToDisplayValues();
         },
 
         getDataToSave = function() {
@@ -304,33 +187,6 @@ exports.init = function(options) {
             };
         }),
 
-        hideAndShowElements = function (fields, actions, action) {
-            if (_.contains(actions, action)) {
-                $(fields).show();
-            } else {
-                if (action === 'edit:start') {
-                    // always hide the applicable runmode buttons
-                    $(fields).filter('.btn').hide();
-
-                    // hide the display fields if there is a corresponding
-                    // edit field to show in its place
-                    _.each($(fields).filter(":not(.btn)"), function (field) {
-                        var $field = $(field),
-                            $edit = FH.getField($(editFields),
-                                                $field.attr('data-field'));
-
-                        if ($edit.length === 1) {
-                            $field.hide();
-                        }
-
-                    });
-
-                } else {
-                    $(fields).hide();
-                }
-            }
-        },
-
         validationErrorsStream = responseErrorStream
             .filter('.fieldErrors')
             .map('.fieldErrors'),
@@ -361,8 +217,6 @@ exports.init = function(options) {
                 return isInEdit ? 'edit:start' : 'cancel';
             });
 
-    $(editFields).find("input[data-date-format]").datepicker();
-
     // Prevent default form submission from clicking on buttons or pressing
     // enter. Event is delegated on window since sometimes <form>s are inserted
     // into the page via AJAX without reiniting inlineEditForm
@@ -377,11 +231,11 @@ exports.init = function(options) {
     actionStream.plug(saveOkStream.map('save:ok'));
     actionStream.plug(responseErrorStream.map('save:error'));
     actionStream.plug(modeChangeStream);
-    actionStream.onValue(hideAndShowElements, editFields, eventsLandingInEditMode);
-    actionStream.onValue(hideAndShowElements, displayFields, eventsLandingInDisplayMode);
-    actionStream.onValue(hideAndShowElements, validationFields, ['save:error']);
+    actionStream.onValue(editForm.hideAndShowElements, editFields, eventsLandingInEditMode);
+    actionStream.onValue(editForm.hideAndShowElements, displayFields, eventsLandingInDisplayMode);
+    actionStream.onValue(editForm.hideAndShowElements, validationFields, ['save:error']);
 
-    saveOKFormDataStream.onValue(formFieldsToDisplayValues);
+    saveOKFormDataStream.onValue(editForm.formFieldsToDisplayValues);
 
     globalErrorsStream.onValue(showGlobalErrors);
     validationErrorsStream.onValue(showValidationErrorsInline);
@@ -389,7 +243,7 @@ exports.init = function(options) {
     unhandledErrorStream.onValue(errorCallback);
     unhandledErrorStream.onValue(logError);
 
-    editStartStream.onValue(displayValuesToFormFields);
+    editStartStream.onValue(editForm.displayValuesToFormFields);
     editStartStream.onValue(showCollectionUdfs);
 
     eventsLandingInDisplayModeStream.onValue(resetCollectionUdfs);
